+++ svgedit-config-iife.js
@@ -16062,7 +16062,7 @@
               }
             }
 
-            return;
+            break;
 
           case 'zoom':
             if (rubberBox != null) {
@@ -16077,7 +16077,7 @@
               height: Math.abs(realY - rStartY),
               factor: factor
             });
-            return;
+            break;
 
           case 'fhpath':
             // Check that the path contains at least 2 points; a degenerate one-point path
@@ -31093,6 +31093,12 @@
             // Add given events to button
             $$b.each(btn.events, function (name, func) {
               if (name === 'click' && btn.type === 'mode') {
+                // `touch.js` changes `touchstart` to `mousedown`,
+                //   so we must map extension click events as well
+                if (isTouch() && name === 'click') {
+                  name = 'mousedown';
+                }
+
                 if (btn.includeWith) {
                   button.bind(name, func);
                 } else {
@@ -33473,6 +33479,8 @@
 
 
               if (opts.evt) {
+                // `touch.js` changes `touchstart` to `mousedown`,
+                //   so we must map tool button click events as well
                 if (isTouch() && opts.evt === 'click') {
                   opts.evt = 'mousedown';
                 }